Code for translation and scale subgizmos is cleaned up
Removed a lot of duplicate code from translation and scale subgizmos.
Common code is now used for both picking and drawing vector- and plane-based subgizmos.
Lints are now defined in Cargo.toml instead of Cranky.toml.
