+++ Cargo.toml
@@ -24,3 +24,156 @@ mint = "0.5"
 [profile.release]
 opt-level = "s"
 lto = "thin"
+
+[lints]
+workspace = true
+
+[workspace.lints.rust]
+elided_lifetimes_in_paths = "allow"
+future_incompatible = "deny"
+nonstandard_style = "deny"
+rust_2018_idioms = "warn"
+rust_2021_prelude_collisions = "deny"
+semicolon_in_expressions_from_macros = "deny"
+trivial_numeric_casts = "deny"
+unsafe_op_in_unsafe_fn = "deny"
+unused_extern_crates = "deny"
+unused_import_braces = "deny"
+unused_lifetimes = "deny"
+unsafe_code = "deny"
+
+[workspace.lints.clippy]
+all = "deny"
+as_ptr_cast_mut = "deny"
+await_holding_lock = "deny"
+bool_to_int_with_if = "deny"
+branches_sharing_code = "deny"
+char_lit_as_u8 = "deny"
+checked_conversions = "deny"
+clear_with_drain = "deny"
+cloned_instead_of_copied = "deny"
+dbg_macro = "deny"
+debug_assert_with_mut_call = "deny"
+default_union_representation = "deny"
+derive_partial_eq_without_eq = "deny"
+disallowed_macros = "deny"
+disallowed_methods = "deny"
+disallowed_names = "deny"
+disallowed_script_idents = "deny"
+disallowed_types = "deny"
+doc_link_with_quotes = "deny"
+doc_markdown = "deny"
+empty_enum = "deny"
+empty_line_after_outer_attr = "deny"
+enum_glob_use = "deny"
+equatable_if_let = "deny"
+exit = "deny"
+expl_impl_clone_on_copy = "deny"
+explicit_deref_methods = "deny"
+explicit_into_iter_loop = "deny"
+explicit_iter_loop = "deny"
+fallible_impl_from = "deny"
+filter_map_next = "deny"
+flat_map_option = "deny"
+float_cmp_const = "deny"
+fn_params_excessive_bools = "deny"
+fn_to_numeric_cast_any = "deny"
+from_iter_instead_of_collect = "deny"
+get_unwrap = "deny"
+if_let_mutex = "deny"
+implicit_clone = "deny"
+imprecise_flops = "deny"
+index_refutable_slice = "deny"
+inefficient_to_string = "deny"
+invalid_upcast_comparisons = "deny"
+iter_not_returning_iterator = "deny"
+iter_on_empty_collections = "deny"
+iter_on_single_items = "deny"
+large_digit_groups = "deny"
+large_include_file = "deny"
+large_stack_arrays = "deny"
+large_stack_frames = "deny"
+large_types_passed_by_value = "deny"
+let_unit_value = "deny"
+linkedlist = "deny"
+lossy_float_literal = "deny"
+macro_use_imports = "deny"
+manual_assert = "deny"
+manual_clamp = "deny"
+manual_instant_elapsed = "deny"
+manual_let_else = "deny"
+manual_ok_or = "deny"
+manual_string_new = "deny"
+map_err_ignore = "deny"
+map_flatten = "deny"
+map_unwrap_or = "deny"
+match_on_vec_items = "deny"
+match_same_arms = "deny"
+match_wild_err_arm = "deny"
+match_wildcard_for_single_variants = "deny"
+mem_forget = "deny"
+mismatched_target_os = "deny"
+mismatching_type_param_order = "deny"
+missing_enforced_import_renames = "deny"
+missing_safety_doc = "deny"
+mut_mut = "deny"
+mutex_integer = "deny"
+needless_borrow = "deny"
+needless_continue = "deny"
+needless_for_each = "deny"
+needless_pass_by_value = "deny"
+negative_feature_names = "deny"
+nonstandard_macro_braces = "deny"
+option_option = "deny"
+path_buf_push_overwrite = "deny"
+print_stdout = "deny"
+ptr_as_ptr = "deny"
+ptr_cast_constness = "deny"
+pub_without_shorthand = "deny"
+rc_mutex = "deny"
+redundant_type_annotations = "deny"
+ref_option_ref = "deny"
+rest_pat_in_fully_bound_structs = "deny"
+same_functions_in_if_condition = "deny"
+semicolon_if_nothing_returned = "deny"
+significant_drop_tightening = "deny"
+single_match_else = "deny"
+str_to_string = "deny"
+string_add_assign = "deny"
+string_add = "deny"
+string_lit_as_bytes = "deny"
+string_to_string = "deny"
+suspicious_command_arg_space = "deny"
+suspicious_xor_used_as_pow = "deny"
+todo = "deny"
+trailing_empty_array = "deny"
+trait_duplication_in_bounds = "deny"
+transmute_ptr_to_ptr = "deny"
+tuple_array_conversions = "deny"
+unchecked_duration_subtraction = "deny"
+unimplemented = "deny"
+uninlined_format_args = "deny"
+unnecessary_box_returns = "deny"
+unnecessary_safety_comment = "deny"
+unnecessary_safety_doc = "deny"
+unnecessary_self_imports = "deny"
+unnecessary_struct_initialization = "deny"
+unnecessary_wraps = "deny"
+unnested_or_patterns = "deny"
+unused_peekable = "deny"
+unused_rounding = "deny"
+unused_self = "deny"
+use_self = "deny"
+useless_transmute = "deny"
+verbose_file_reads = "deny"
+wildcard_dependencies = "deny"
+zero_sized_map_values = "deny"
+let_underscore_untyped = "allow"
+missing_assert_message = "allow"
+missing_errors_doc = "allow"
+too_many_lines = "allow"
+undocumented_unsafe_blocks = "allow"
+unwrap_used = "allow"
+wildcard_imports = "allow"
+large-types-passed-by-value = "allow"
+needless-pass-by-value = "allow"
