+++ rust-toolchain
@@ -1,4 +1,4 @@
 [toolchain]
-channel = "1.72.0"
+channel = "1.74.0"
 components = ["rustfmt", "clippy"]
 targets = ["wasm32-unknown-unknown"]
